requirements.txt
fitz
pytesseract
pdf2image
camelot-py[cv]
pandas
pdfplumber
tabula-py
requests
